Added custom graph support, and cleaned up code and UI

+++ src/com/functiongrapher/function/Function.java
@@ -6,6 +6,6 @@ public interface Function {
 	
 	public int getDrawMode();
 	
-	public byte[] getGraphColor();
+	public double[] getGraphColor();
 	
 }
